refactor(types): move MatchMachineContext to src/game/types/index.ts
